+++ test/bfa8.jl
@@ -1,11 +1,11 @@
-@testset "BFARadPowerModel" begin
+@testset "BFA8PowerModel" begin
 
     @testset "CIGRE TNEP single-period" begin
         data = _FP.parse_file(normpath(@__DIR__,"../test/data/cigre_mv_eu/cigre_mv_eu_unit_test.m"))
         _FP.add_dimension!(data, :hour, 1)
         _FP.add_dimension!(data, :year, 1)
         data = _FP.make_multinetwork(data)
-        result = _FP.flex_tnep(data, _FP.BFARadPowerModel, milp_optimizer)
+        result = _FP.flex_tnep(data, _FP.BFA8PowerModel, milp_optimizer)
         sol = result["solution"]["nw"]["1"]
         @test result["termination_status"] == OPTIMAL
         @test result["objective"] ≈ 4360.45 rtol=1e-3
@@ -30,7 +30,7 @@
         data["nw"]["1"]["load"]["1"]["pd"] += 10.0 # Bus 1. Changes reactive power demand too, via `pf_angle`.
         data["nw"]["1"]["load"]["12"]["pd"] += 4.0 # Bus 13. Changes reactive power demand too, via `pf_angle`.
         data["nw"]["1"]["branch"]["12"]["rate_a"] = data["nw"]["1"]["branch"]["12"]["rate_b"] = data["nw"]["1"]["branch"]["12"]["rate_c"] = 0.0
-        result = _FP.flex_tnep(data, _FP.BFARadPowerModel, milp_optimizer)
+        result = _FP.flex_tnep(data, _FP.BFA8PowerModel, milp_optimizer)
         sol = result["solution"]["nw"]["1"]
         @test result["termination_status"] == OPTIMAL
         @test result["objective"] ≈ 5764.48 rtol=1e-3