+++ test/storage.jl
@@ -1,6 +1,6 @@
 # Test storage model using a multiperiod optimization
 
-# - Model: `_FP.BFARadPowerModel` is used to be able to test storage in both active and
+# - Model: `_FP.BFA8PowerModel` is used to be able to test storage in both active and
 #   reactive power, while keeping the model linear. It requires a distribution network.
 # - Problem: `flex_tnep` is the simplest problem that implements both storage and flexible
 #   loads.
@@ -70,7 +70,7 @@ end
         loadprofile = collect(reshape(range(0,2;length=number_of_hours),:,1)) # Create a load profile: ramp from 0 to 2 times the rated value of load
         time_series = _FP.make_time_series(data; loadprofile) # Compute time series by multiplying the rated value by the profile
         mn_data = _FP.make_multinetwork(data, time_series)
-        result = _FP.flex_tnep(mn_data, _FP.BFARadPowerModel, milp_optimizer)
+        result = _FP.flex_tnep(mn_data, _FP.BFA8PowerModel, milp_optimizer)
 
         @testset "Existing storage" begin
             #plot_storage(mn_data, result; candidate=false)
@@ -130,7 +130,7 @@ end
         loadprofile = collect(reshape(range(0,2;length=number_of_hours),:,1)) # Create a load profile: ramp from 0 to 2 times the rated value of load
         time_series = _FP.make_time_series(data; loadprofile) # Compute time series by multiplying the rated value by the profile
         mn_data = _FP.make_multinetwork(data, time_series)
-        result = _FP.flex_tnep(mn_data, _FP.BFARadPowerModel, milp_optimizer)
+        result = _FP.flex_tnep(mn_data, _FP.BFA8PowerModel, milp_optimizer)
 
         @testset "Existing storage" begin
             #plot_storage(mn_data, result; candidate=false)
@@ -165,7 +165,7 @@ end
         loadprofile = collect(reshape(range(0,1.1005;length=number_of_hours),:,1)) # Create a load profile: ramp from 0 to 1.1005 times the rated value of load
         time_series = _FP.make_time_series(data; loadprofile) # Compute time series by multiplying the rated value by the profile
         mn_data = _FP.make_multinetwork(data, time_series)
-        result = _FP.flex_tnep(mn_data, _FP.BFARadPowerModel, milp_optimizer)
+        result = _FP.flex_tnep(mn_data, _FP.BFA8PowerModel, milp_optimizer)
 
         @testset "Not built if not needed" begin
             #plot_storage(mn_data, result; candidate=true)